Apply manual updaates to fix rubocop offenses
