+++ lib/cucumber/events/step_definition_registered.rb
@@ -14,6 +14,8 @@ module Cucumber
       # _@private
       def initialize(step_definition)
         @step_definition = step_definition
+
+        super
       end
     end
   end
