+++ lib/cucumber/formatter/ignore_missing_messages.rb
@@ -11,9 +11,11 @@ module Cucumber
         @receiver.respond_to?(message) ? @receiver.send(message, *args) : super
       end
 
+      # rubocop:disable Lint/MissingSuper
       def respond_to_missing?(name, include_private = false)
         @receiver.respond_to?(name, include_private)
       end
+      # rubocop:enable Lint/MissingSuper
     end
   end
 end
